--- a/.gitmodules
+++ b/.gitmodules
@@ -2,3 +2,7 @@
 	path = lib/forge-std
 	url = https://github.com/foundry-rs/forge-std
 	branch = v1.3.0
+[submodule "tests/integration/ef_tests"]
+	path = tests/integration/ef_tests/testdata
+	url = https://github.com/ethereum/tests
+  shallow = true