--- a/.gitmodules
+++ b/.gitmodules
@@ -3,6 +3,6 @@
 	url = https://github.com/foundry-rs/forge-std
 	branch = v1.3.0
 [submodule "tests/integration/ef_tests"]
-	path = tests/integration/ef_tests/testdata
+	path = tests/integration/ef_tests/test_data
 	url = https://github.com/ethereum/tests
-  shallow = true
+	shallow = true