--- a/tests/integration/ef_tests/conftest.py
+++ b/tests/integration/ef_tests/conftest.py
@@ -1,74 +1,9 @@
-import logging
-
-from utils import (
-    filter_by_case_ids,
-    load_default_ef_blockchain_tests,
-    load_ef_blockchain_tests,
-)
-
-logger = logging.getLogger()
+import pytest
 
 
-def pytest_addoption(parser):
-    parser.addoption(
-        "--target",
-        action="store",
-        type=str,
-        help="Specify ef test directory or suite to run, where a suite is a json file.",
-        default=None,
-    )
-    parser.addoption(
-        "--case",
-        action="append",
-        type=str,
-        nargs="+",
-        help="Specify one or more particular cases within the JSON suite files to run",
-        default=None,
-    )
-    parser.addoption(
-        "--network",
-        action="store",
-        type=str,
-        default="Shanghai",
-        help="Specify the network to use, defaults to Shanghai",
-    )
+def pytest_runtest_setup(item):
+    keywordexpr = item.config.getoption("keyword")
 
-
-def pytest_generate_tests(metafunc):
-    if "ef_blockchain_test" in metafunc.fixturenames:
-        suite_or_directory = metafunc.config.getoption("target")
-        case_ids = metafunc.config.getoption("case")
-        network_name = metafunc.config.getoption("network")
-
-        # Initialize as empty lists
-        ef_blockchain_test_ids = []
-        ef_blockchain_test_objects = []
-
-        if suite_or_directory:
-            ef_blockchain_tests = load_ef_blockchain_tests(
-                suite_or_directory, network_name
-            )
-        else:
-            ef_blockchain_tests = load_default_ef_blockchain_tests(network_name)
-
-        if case_ids:
-            flattened_case_ids = [item for sublist in case_ids for item in sublist]
-            ef_blockchain_tests = filter_by_case_ids(
-                ef_blockchain_tests, flattened_case_ids
-            )
-
-        if not ef_blockchain_tests:
-            logger.warning(
-                f"No tests found for `--target` param {suite_or_directory} and `--case_ids` {case_ids}. Skipping tests."
-            )
-        else:
-            ef_blockchain_test_ids, ef_blockchain_test_objects = zip(
-                *ef_blockchain_tests
-            )
-
-        # Parametrize regardless of whether ef_blockchain_tests is empty or not
-        metafunc.parametrize(
-            "ef_blockchain_test",
-            ef_blockchain_test_objects,
-            ids=ef_blockchain_test_ids,
-        )
+    # If the keyword expression doesn't match the nodeid (test name), skip the test
+    if item.nodeid.find(keywordexpr) == -1 or not keywordexpr :
+        pytest.skip("Skipping test, didn't match keyword expression.")