--- a/tests/integration/ef_tests/conftest.py
+++ b/tests/integration/ef_tests/conftest.py
@@ -1,6 +1,14 @@
+import logging
 import time
 
-from utils import load_all_ef_blockchain_tests
+from utils import (
+    is_directory_or_file_keyword,
+    load_all_ef_blockchain_tests,
+    load_ef_blockchain_tests,
+)
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
 
 
 # This enables devs to use the familiar keyword (-k argument) to select a subset of the EF test suite,
@@ -8,7 +16,7 @@ from utils import load_all_ef_blockchain_tests
 def pytest_generate_tests(metafunc):
     if "ef_blockchain_test" in metafunc.fixturenames:
         keyword = metafunc.config.getoption("keyword")
-
+        logger.info(f"EF test runner ran with {keyword=}")
         if not keyword:
             # with no keyword, we parametrize as empty
             metafunc.parametrize(
@@ -16,9 +24,20 @@ def pytest_generate_tests(metafunc):
                 [],
                 ids=[],
             )
+
         else:
             start_time = time.time()
-            ef_blockchain_tests = load_all_ef_blockchain_tests()
+
+            if is_directory_or_file_keyword(keyword):
+                # allows user to not load all ef-tests to then filter
+                # a user can either load all tests in a subdirectory or in a json file
+                # following the directory structure here:
+                # https://github.com/ethereum/tests/tree/develop/BlockchainTests/GeneralStateTests
+
+                ef_blockchain_tests = load_ef_blockchain_tests(keyword)
+                metafunc.config.option.keyword = ""  # Override keyword expression to avoid deselection when using a 'direct' keyword.
+            else:
+                ef_blockchain_tests = load_all_ef_blockchain_tests()
             # Calculate and print the elapsed time
             elapsed_time = time.time() - start_time
             print(f"Time taken to load_ef_blockchain_tests: {elapsed_time} seconds")