<!--- Please provide a general summary of your changes in the title above -->

<!-- Give an estimate of the time you spent on this PR in terms of work days.
Did you spend 0.5 days on this PR or rather 2 days?  -->

Time spent on this PR: 0.1 

## Pull request type

<!-- Please try to limit your pull request to one type,
submit multiple pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [X] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Building off #750, we add a particular usage of the -k param that allows a dev to load either a directory or file, as opposed to using the keyword to select over all the tests loaded in memory.

Motivating example

Ethereum foundation tests have a directory structure, for example https://github.com/ethereum/tests/tree/develop/BlockchainTests/GeneralStateTests/VMTests

To attempt to load just some of the suites in this directory structure:

```
poetry run pytest ./tests/integration/ef_tests/ -k "mulmod or swap or calldatasize or push or random or calldatacopy" -vvv -rA

collecting ... Time taken to load_ef_blockchain_tests: 4.282881021499634 seconds

collecting 19721 items                                                                                                   
collected 19721 items / 18906 deselected / 815 selected                                                                  

```

to run a particular suite

```
poetry run pytest ./tests/integration/ef_tests/ -k "sha3" -vvv -rA

collecting ... Time taken to load_ef_blockchain_tests: 4.460179328918457 seconds

collecting 19721 items                                                                                                   
collected 19721 items / 19700 deselected / 21 selected                                                                   

```



## What is the new behavior?

Allow a user to just use  a directory/file input, where we get a more refined selection at a faster speed

```
poetry run pytest ./tests/integration/ef_tests/ -k "VMTests/" -vvv -rA

collecting ... Time taken to load_ef_blockchain_tests: 0.2358388900756836 seconds

collected 651 items                                                                                                      
```

```
poetry run pytest ./tests/integration/ef_tests/ -k "VMTests/vmTests/sha3.json" -vvv -rA

collecting ... Time taken to load_ef_blockchain_tests: 0.00374603271484375 seconds

collected 17 items
```

Note how not only is this process slower, but also the keyword filtering overmatches and has tests that are outside the scope of particular suites, which is what would be one of the core units of reasoning for debugging work.